Time: 110 ms (28.99%), Space: 57.2 MB (78.51%) - LeetHub

+++ 0084-largest-rectangle-in-histogram/0084-largest-rectangle-in-histogram.java
@@ -24,11 +24,11 @@ class Solution {
             else pse[i]=st.peek();
             st.push(i);
         }
-        long rectangle=-1;
+        int rectangle=-1;
         for(int i=0;i<n;i++)
         {
             rectangle=Math.max(rectangle,(heights[i]*(nse[i]-pse[i]-1)));
         }
-        return (int)rectangle;
+        return rectangle;
     }
 }
